Android 17.0.0 release 1

* tag 'android-17.0.0_r1': (1187 commits)
  Disable hard inode quotas by default.
  Restrict inode quota setup to application UIDs.
  Disable Kawase2's blur texture preallocation for unprotected on GaneshGL
  vulkan: fix to advertise EXT_hdr_metadata unconditionally
  SF: null deref fix
  Expect that IAccessor is SYSTEM stability
  Move internal color type into GaneshBackendTexture
  Fix SensorPrivacyManager support map cache poisoning
  Actually let the fake dataspace do an awkward cast in renderengine.
  Revert^2 "Wait for boot complete before spinning up aisealhostservice"
  Add support for appId <= 0 for installd#getAppSize
  Revert "Wait for boot complete before spinning up aisealhostservice"
  Use Skia's new HLG transfer definition for SMPTE 2094-50
  Misc. memory debugging improvements
  Revert "inputflinger: Use a single WindowInfosListener"
  Revert "PointerChoreographer: use mLock directly"
  Revert "inputflinger: Use a single WindowInfosListener"
  Revert "PointerChoreographer: use mLock directly"
  Wait for boot complete before spinning up aisealhostservice
  renderengine: Pass color matrix to BoxShadows and Border
  ...

+++ services/inputflinger/reader/mapper/RotaryEncoderInputMapper.h
@@ -55,7 +55,7 @@ private:
      * the IDC file, or if a non-positive value is provided, the telemetry will be disabled, and
      * this value is set to the empty optional.
      */
-    std::optional<int32_t> mMinRotationsToLog;
+    std::optional<float> mMinRotationsToLog;
     /**
      * A function to log count for telemetry.
      * The char* is the logging key, and the int64_t is the value to log.
